+++ src/containers/Home/index.js
@@ -7,7 +7,7 @@ import { connect } from 'react-redux';
 
 import debug from '../../helpers/debug';
 
-import { isAllLoaded, loadAll } from '../../redux/modules/surahs';
+import { isAllLoaded, loadAll } from '../../redux/actions/SurahsActions.js';
 
 const styles = require('./style.scss');
