+++ src/containers/Surah/connect.js
@@ -1,4 +1,4 @@
-import { isAllLoaded, loadAll, setCurrent as setCurrentSurah } from '../../redux/modules/surahs';
+import { isAllLoaded, loadAll, setCurrent as setCurrentSurah } from '../../redux/actions/SurahsActions.js';
 import { clearCurrent, isLoaded, load as loadAyahs } from '../../redux/actions/AyahsActions.js';
 
 import debug from 'helpers/debug';
