+++ src/redux/constants/SurahsActionTypes.js
@@ -1,0 +1,7 @@
+export const LOAD = '@@quran/surahs/LOAD';
+export const LOAD_SUCCESS = '@@quran/surahs/LOAD_SUCCESS';
+export const LOAD_FAIL = '@@quran/surahs/LOAD_FAIL';
+export const LOAD_INFO = '@@quran/surahs/LOAD_INFO';
+export const LOAD_INFO_SUCCESS = '@@quran/surahs/LOAD_INFO_SUCCESS';
+export const LOAD_INFO_FAIL = '@@quran/surahs/LOAD_INFO_FAIL';
+export const SET_CURRENT = '@@quran/surahs/SET_CURRENT';
